Update ttest2w.m
fixed error in calculation of weighted variance. I figured out both weighted mean and variance can be calculated by Matlab's built-in variance function. My original variance formula was wrong anyway (the original version was based on the SAS specification which differs from the formula detailed in the documentation for 'var'. I'm no stats expert, so I don't know which one is correct, but I'm just going with Matlab's version and they can take the blame if their formula is wrong.)